chore(ci): refactor fixtures to set storage state deterministically, add reusable storageState file
