+++ apps/extension/tests/utils.ts
@@ -1,7 +1,12 @@
-import { Locator } from '@playwright/test';
+import { expect, Locator, Page } from '@playwright/test';
 import { makeUnsignedSTXTokenTransfer } from '@stacks/transactions';
 
+import { HomePageSelectors } from './selectors/home.selectors';
 import { SharedComponentsSelectors } from './selectors/shared-component.selectors';
+import {
+  TEST_ACCOUNT_DERIVED_KEY,
+  testSoftwareAccountDefaultWalletState,
+} from './fixtures/wallet-state';
 
 export function json(arg: unknown) {
   return {
@@ -65,4 +70,26 @@ export async function generateMultisigUnsignedStxTransfer(
 // Intl formatters like NumberFormat output non-breaking spaces. This helper can wrap expected strings to avoid manually writing \u00A0 in tests.
 export function withNbsp(value: string) {
   return value.replace(/ /g, '\u00A0');
+}
+
+export async function seedSoftwareWallet(page: Page, extensionId: string) {
+  await page.goto(`chrome-extension://${extensionId}/index.html`);
+  await page.waitForLoadState('domcontentloaded');
+  await page.waitForFunction(() => Boolean(window.chrome?.storage?.local));
+
+  const walletState = JSON.parse(JSON.stringify(testSoftwareAccountDefaultWalletState));
+
+  await page.evaluate(
+    async ({ state, encryptionKey }) => {
+      await Promise.all([
+        new Promise<void>(resolve => chrome.storage.local.set({ 'persist:root': state }, () => resolve())),
+        new Promise<void>(resolve => chrome.storage.session.set({ encryptionKey }, () => resolve())),
+      ]);
+    },
+    { state: walletState, encryptionKey: TEST_ACCOUNT_DERIVED_KEY }
+  );
+
+  await page.reload();
+  await page.waitForLoadState('domcontentloaded');
+  await expect(page.getByTestId(HomePageSelectors.HomePageContainer)).toBeVisible();
 }
